backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
data==0.4
dataclasses-json==0.5.9
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.3.0
exceptiongroup==1.1.1
fastapi==0.99.0
fpdf==1.7.2
frozenlist==1.3.3
funcsigs==1.0.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
langchain<=0.0.220
loguru==0.7.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy<=1.25.0
openapi-schema-pydantic<=1.2.4
packaging<=23.1
pinecone-client<=2.2.2
pycparser<=2.21
pydantic<=1.10.10
PyGithub<=1.59.0
PyJWT<=2.7.0
PyNaCl==1.5.0
python-dateutil<=2.8.2
python-dotenv<=1.0.0
PyYAML<=6.0
requests<=2.31.0
six<=1.16.0
sniffio<=1.3.0
SQLAlchemy<=2.0.17
starlette<=0.27.0
tenacity<=8.2.2
tqdm<=4.65.0
typing-inspect<=0.9.0
typing_extensions<=4.7.0
urllib3<=2.0.3
uvicorn<=0.22.0
wrapt<=1.15.0
yarl<=1.9.2
urllib3==1.26.6
openai<=0.27.8
pypdf<=3.11.1
tiktoken<=0.4.0
chromadb<=0.3.26